Moved account settings to its own page and linked it from the settings module
