+++ src/app/membership/account/account-routing.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+
+import { AccountPage } from './account.page';
+
+const routes: Routes = [
+  {
+    path: '',
+    component: AccountPage
+  }
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule],
+})
+export class AccountPageRoutingModule {}
