+++ src/app/membership/account/account.module.ts
@@ -1,0 +1,20 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { ReactiveFormsModule } from '@angular/forms';
+
+import { IonicModule } from '@ionic/angular';
+
+import { AccountPageRoutingModule } from './account-routing.module';
+
+import { AccountPage } from './account.page';
+
+@NgModule({
+  imports: [
+    CommonModule,
+    ReactiveFormsModule,
+    IonicModule,
+    AccountPageRoutingModule
+  ],
+  declarations: [AccountPage]
+})
+export class AccountPageModule {}
