+++ src/app/membership/account/account.page.html
@@ -1,0 +1,125 @@
+<ion-header>
+  <ion-toolbar>
+    <ion-buttons slot="start">
+      <ion-back-button></ion-back-button>
+    </ion-buttons>
+    <ion-title>Account Settings</ion-title>
+  </ion-toolbar>
+</ion-header>
+
+<ion-content>
+
+  <ion-grid>
+    <ion-col size-md="6" size-lg="5" size-xs="12">
+      <form [formGroup]="emailForm">
+
+        <div class="ion-text-center">
+          <h3>Account Details</h3>
+        </div>
+
+        <ion-item-divider mode="md"></ion-item-divider>
+
+        <ion-item>
+          <ion-icon slot="start" name="mail"></ion-icon>
+          <ion-label *ngIf="!editEmail">{{user.getEmail()}}</ion-label>
+          <ion-input *ngIf="editEmail" formControlName="email" type="email" placeholder="New Email"></ion-input>
+          <ion-button slot="end" *ngIf="edit && !editUsername && !editPassword" (click)="canEditEmail()">
+            <ion-icon slot="icon-only" size="small" [name]="editEmail ? 'lock-open':'lock-closed'"></ion-icon>
+          </ion-button>
+
+        </ion-item>
+        <div *ngFor="let validation of validationMessages['email']">
+          <ion-item lines="none" *ngIf="emailForm.controls.email.hasError(validation['type']) && emailForm.controls.email.touched && editEmail">
+            <ion-text color="danger">
+              <sub>*{{ validation['message'] }}</sub>
+            </ion-text>
+          </ion-item>
+        </div>
+      </form>
+      <ion-button *ngIf="editEmail"
+                  type="submit"
+                  expand="block"
+                  [disabled]="emailForm.invalid || buttonPressed"
+                  (click)="updateEmail(emailForm)">
+        Update Email
+      </ion-button>
+      <form [formGroup]="usernameForm">
+        <ion-item>
+          <ion-icon slot="start" name="person"></ion-icon>
+          <ion-label *ngIf="!editUsername">{{user.getUsername()}}</ion-label>
+          <ion-input *ngIf="editUsername" formControlName="username" placeholder="New Username"></ion-input>
+          <ion-button slot="end" *ngIf="edit && !editEmail && !editPassword" (click)="canEditUsername()">
+            <ion-icon slot="icon-only" size="small" [name]="editUsername ? 'lock-open':'lock-closed'"></ion-icon>
+          </ion-button>
+        </ion-item>
+        <div *ngFor="let validation of validationMessages['username']">
+          <ion-item lines="none" *ngIf="usernameForm.controls.username.hasError(validation['type']) && usernameForm.controls.username.touched && editUsername">
+            <ion-text color="danger">
+              <sub>*{{ validation['message'] }}</sub>
+            </ion-text>
+          </ion-item>
+        </div>
+      </form>
+      <ion-button *ngIf="editUsername"
+                  type="submit"
+                  expand="block"
+                  [disabled]="usernameForm.invalid || buttonPressed"
+                  (click)="updateUsername(usernameForm)">
+        Update Username
+      </ion-button>
+      <form [formGroup]="passwordForm">
+        <ion-item>
+          <ion-icon slot="start" name="lock-closed"></ion-icon>
+          <ion-label *ngIf="!editPassword">****************</ion-label>
+          <ion-input *ngIf="editPassword" formControlName="old" type="password" placeholder="Old Password"></ion-input>
+          <ion-button slot="end" *ngIf="edit && !editEmail && !editUsername" (click)="canEditPassword()">
+            <ion-icon slot="icon-only" size="small" [name]="editPassword ? 'lock-open':'lock-closed'"></ion-icon>
+          </ion-button>
+        </ion-item>
+        <div *ngFor="let validation of validationMessages['password']">
+          <ion-item lines="none" *ngIf="passwordForm.controls.old.hasError(validation['type']) && passwordForm.controls.old.touched">
+            <ion-text color="danger">
+              <sub>*{{ validation['message'] }}</sub>
+            </ion-text>
+          </ion-item>
+        </div>
+        <div formGroupName="password">
+          <ion-item *ngIf="editPassword">
+            <ion-icon slot="start" name="key-outline"></ion-icon>
+            <ion-input formControlName="set" type="password" placeholder="New Password"></ion-input>
+          </ion-item>
+          <ion-item *ngIf="editPassword">
+            <ion-icon slot="start" name="key"></ion-icon>
+            <ion-input formControlName="confirm" type="password" placeholder="Confirm New Password"></ion-input>
+          </ion-item>
+        </div>
+        <div *ngFor="let validation of validationMessages['password']">
+          <ion-item lines="none" *ngIf="passwordForm.controls.password.get('set').hasError(validation['type']) && passwordForm.controls.password.get('set').touched">
+            <ion-text color="danger">
+              <sub>*{{ validation['message'] }}</sub>
+            </ion-text>
+          </ion-item>
+        </div>
+        <ion-item lines="none" *ngIf="passwordForm.controls.password.errors !== null && passwordForm.controls.password.get('confirm').touched">
+          <ion-text color="danger">
+            <sub>*{{ validationMessages['confirm']['message']}}</sub>
+          </ion-text>
+        </ion-item>
+      </form>
+      <ion-button *ngIf="editPassword"
+                  type="submit"
+                  expand="block"
+                  [disabled]="passwordForm.invalid || buttonPressed"
+                  (click)="updatePassword(passwordForm)">
+        Update Password
+      </ion-button>
+      <ion-row class="ion-justify-content-end">
+        <ion-button size="small" color="dark" (click)="toggleEdit()">
+          <ion-label>Edit</ion-label>
+          <ion-icon slot="end" size="small" [name]="edit ? 'lock-open':'lock-closed'"></ion-icon>
+        </ion-button>
+      </ion-row>
+    </ion-col>
+  </ion-grid>
+
+</ion-content>
