+++ src/app/membership/account/account.page.ts
@@ -1,0 +1,129 @@
+import { Component, OnInit } from '@angular/core';
+import { FormGroup } from '@angular/forms';
+
+import { FormService } from '../../global/services/form.service';
+import { LoadingService } from '../../global/services/loading.service';
+import {ToastService} from '../../global/services/toast.service';
+
+import {UserService} from '../authentication/user.service';
+
+@Component({
+  selector: 'app-account',
+  templateUrl: './account.page.html',
+  styleUrls: ['./account.page.scss'],
+})
+export class AccountPage implements OnInit {
+
+  private emailForm: FormGroup;
+  private usernameForm: FormGroup;
+  private passwordForm: FormGroup;
+  private validationMessages: any;
+  private edit: boolean;
+  private editEmail: boolean;
+  private editUsername: boolean;
+  private editPassword: boolean;
+  private buttonPressed: boolean;
+
+  constructor(
+    private loading: LoadingService,
+    private toast: ToastService,
+    private user: UserService
+  ) {
+    this.emailForm = FormService.emailForm();
+    this.usernameForm = FormService.usernameForm();
+    this.passwordForm = FormService.passwordForm();
+    this.validationMessages = FormService.validationMessages();
+    this.edit = false;
+    this.editEmail = false;
+    this.editUsername = false;
+    this.editPassword = false;
+    this.buttonPressed = false;
+  }
+
+  ngOnInit() {
+  }
+
+  toggleEdit() {
+    if (this.edit) { this.resetButtons(); } else { this.edit = true; }
+  }
+
+  canEditEmail() {
+    this.editEmail = !this.editEmail;
+  }
+
+  canEditUsername() {
+    this.editUsername = !this.editUsername;
+  }
+
+  canEditPassword() {
+    this.editPassword = !this.editPassword;
+  }
+
+  resetButtons() {
+    this.edit = this.editEmail = this.editUsername = this.editPassword = this.buttonPressed = false;
+  }
+
+  async updateEmail(form) {
+    this.buttonPressed = true;
+    await this.loading.getLoading('Updating Email...');
+    this.user.setEmail(form.value).subscribe( {
+      next: async res => {
+        console.log(res);
+        this.loading.dismiss().then(() => {
+          this.toast.showSuccess('Email has been updated to ' + res.email);
+          this.resetButtons();
+        });
+      },
+      error: async err => {
+        console.log(err.status);
+        this.loading.dismiss().then(() => {
+          this.toast.showError(err.status);
+          this.buttonPressed = false;
+        });
+      }
+    });
+  }
+
+  async updateUsername(form) {
+    this.buttonPressed = true;
+    await this.loading.getLoading('Updating Username...');
+    this.user.setUsername(form.value).subscribe({
+      next: async res => {
+        console.log(res);
+        this.loading.dismiss().then(() => {
+          this.toast.showSuccess('Username has been updated to ' + res.username);
+          this.resetButtons();
+        });
+      },
+      error: async err => {
+        console.log(err.status);
+        this.loading.dismiss().then(() => {
+          this.toast.showError(err.status);
+          this.buttonPressed = false;
+        });
+      }
+    });
+  }
+
+  async updatePassword(form) {
+    this.buttonPressed = true;
+    await this.loading.getLoading('Updating Password');
+    this.user.setPassword(form.value).subscribe({
+      next: async res => {
+        console.log(res);
+        this.loading.dismiss().then(() => {
+          this.toast.showSuccess('Password has been successfully updated!');
+          this.resetButtons();
+        });
+      },
+      error: async err => {
+        console.log(err.status);
+        this.loading.dismiss().then(() => {
+          this.toast.showError(err.status);
+          this.buttonPressed = false;
+        });
+      }
+    });
+  }
+
+}
